// src/app/interfaces/beneficiaries-dev-room-list-dto.ts
export interface BeneficiariesDevRoomListDto {
    TrainingCenterId: string;
    TrainingCenterName: string;
    CampusId: string;
    CampusName: string;
    DevelopmentRoomId: string;
    DevelopmentRoomName: string;
    GroupId: string;
    GroupName: string;
    Year: string;
    Beneficiary: string;
}